+++ server/api/v1/services/collection.service.ts
@@ -1,15 +1,14 @@
 import jwt from "jsonwebtoken";
-import mongoose from "mongoose";
 import { ObjectId } from "mongoose"
 
 import { UserModel } from "../models/user.model";
 import { ClusterModel } from "../models/cluster.model";
 import { CollectionModel } from "../models/collection.model";
 
+import { Collection } from "../interfaces/collection.interface";
+
 import { tryToCatch } from "../utils";
 import { logger } from "../logger";
-import { Collection } from "../interfaces/collection.interface";
-import cluster, { Cluster } from "cluster";
 
 export class CollectionService {
 
@@ -19,8 +18,8 @@ export class CollectionService {
 
   constructor (token: string, id: string | null, collectionInfo: Collection<ObjectId> | null) {
     this.token = token;
-    this.collectionInfo = collectionInfo;
     this.id = id;
+    this.collectionInfo = collectionInfo;
   };
 
   async tokenTransformer() {
@@ -32,10 +31,11 @@ export class CollectionService {
     let [err, data] = await tryToCatch(async (userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     };
+    if (!data) return { error: true, detail: "Unkown user!" };
     
-    [err, data] = await tryToCatch(async () => CollectionModel.find({}));
+    [err, data] = await tryToCatch(async (id: string) => CollectionModel.find({clusterParent: id}), data.cluster);
     if (err) {
       logger.error(err);
       return { error: true, detail: err };
@@ -47,13 +47,13 @@ export class CollectionService {
     let [err, data] = await tryToCatch(async (userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     };  
 
     [err, data] = await tryToCatch(async (id: string) => CollectionModel.findById(id), this.id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Empty" };
+      return { error: true, detail: "Not found!" };
     };return { error: false, detail: data };
 
   };
@@ -63,8 +63,9 @@ export class CollectionService {
     let [err, data] = await tryToCatch(async (userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     }; 
+    if (!data) return { error: true, detail: "Unkown user!" };
 
     [err, data] = await tryToCatch(async (id: string) => {
       const collectionDoc = await CollectionModel.findById(id);
@@ -72,7 +73,7 @@ export class CollectionService {
     }, this.id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Collection doesn't exists!" };
+      return { error: true, detail: "Not found!" };
     };return { error: typeof data === "string", detail: data };
   };
 
@@ -81,23 +82,36 @@ export class CollectionService {
     let [err, data] = await tryToCatch(async (userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
-    }; 
+      return { error: true, detail: err };
+    };
+    if (!data) return { error: true, detail: "Unkown user!" };
 
     [err, data] = await tryToCatch(async (id: string) => {
-      const collectionDoc = await CollectionModel.findById(id);
       const clusterDoc = await ClusterModel.findById(data.cluster);
+      const allCollections = await CollectionModel.find({clusterParent: data.cluster});
+      const collectionDoc = await CollectionModel.findById(id).lean();
 
-      for (let coll of clusterDoc.collections) {
-        if (coll.toString() === id) {
-          return { error: true, detail: "You already had this collection!" };
-        };
+      if (data.cluster.toString() === collectionDoc.clusterParent.toString()) 
+        return "You already had this collection!";
+      if (!collectionDoc.shared) 
+        return "This collection is private!";
+
+      const {_id, __v, ...collection} = collectionDoc;
+
+      for (let coll of allCollections) {
+        if (collection.name === coll.name) 
+          collection.name += `(shared ${Math.random().toString(16).substr(2, 4)})`;
       };
 
-      if (!collectionDoc.shared) return { error: false, detail: "This cluster is private!" };
-      clusterDoc.collections.push(collectionDoc);
+      collection.clusterParent = data.cluster;
+
+      const newCollectionDoc = new CollectionModel({...collection});
+      newCollectionDoc.save()
+
+      clusterDoc.collections.push(newCollectionDoc);
       clusterDoc.save();
-      return "Collestion saved!";
+
+      return newCollectionDoc;
     }, this.id);
     if (err) {
       logger.error(err);
@@ -110,8 +124,9 @@ export class CollectionService {
     let [err, data] = await tryToCatch(async (userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     }; 
+    if (!data) return { error: true, detail: "Unkown user!" };
 
     [err, data] = await tryToCatch(async (collection: Collection<ObjectId>) => {
       // check if collection belong to the right user collection      
@@ -142,19 +157,22 @@ export class CollectionService {
     let [err, data] = await tryToCatch((userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     };
+    if (!data) return { error: true, detail: "Unkown user!" };
 
     [err, data] = await tryToCatch(async (update: Collection<ObjectId>) => {
       
       const collectionDoc = await CollectionModel.findOne({_id: this.id, clusterParent: data.cluster});
-      // checifcluster name does exists!      
+      // check if cluster name does exists!      
       const collectionNameDoesExists = await CollectionModel.find(
         {
           clusterParent: data.cluster, name: update.name
         }
       );
-      if (collectionNameDoesExists.length) return "Collection name already exists!";
+      collectionDoc.shared = update.shared
+      await collectionDoc.save();
+      if (collectionNameDoesExists.length) return collectionDoc;
 
       collectionDoc.name = update.name;
       collectionDoc.shared = update.shared
@@ -174,8 +192,9 @@ export class CollectionService {
     let [err, data] = await tryToCatch((userId: string) => UserModel.findById(userId), _id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: "Unkown user!" };
+      return { error: true, detail: err };
     };
+    if (!data) return { error: true, detail: "Unkown user!" };
 
     [err, data] = await tryToCatch(async (id: string) => {
       const res = await CollectionModel.deleteOne({_id: id, clusterParent: data.cluster});
@@ -185,7 +204,7 @@ export class CollectionService {
     }, this.id);
     if (err) {
       logger.error(err);
-      return { error: true, detail: err };
+      return { error: true, detail: "Couldn't delete collection!" };
     };return { error: false, detail: "Collection has been deleted!" };
   };
